Partial OpaquePool implementation

+++ include/ae/aether.h
@@ -1426,7 +1426,7 @@ public:
 	const T* GetLast() const;
 
 	template < typename U > T* Find( const U& value );
-	template < typename Fn > T* FindFn( Fn predicateFn );
+	template < typename Fn > T* FindFn( Fn predicateFn ); // @TODO: FindFn's parameter should be a reference to match ae::Array
 
 	uint32_t Length() const;
 
@@ -6970,7 +6970,7 @@ bool FreeList< N >::IsAllocated( int32_t idx ) const
 		return false;
 	}
 #if _AE_DEBUG_
-	AE_ASSERT( idx < m_pool.Length() );
+	AE_ASSERT( (uint32_t)idx < m_pool.Length() );
 #endif
 	return !m_pool[ idx ].next;
 }
@@ -7014,7 +7014,6 @@ ObjectPool< T, N, Paged >::~ObjectPool()
 template < typename T, uint32_t N, bool Paged >
 T* ObjectPool< T, N, Paged >::New()
 {
-	// @TODO: FindFn's parameter should be a reference to match ae::Array
 	Page* page = m_pages.FindFn( []( const Page* page ) { return page->freeList.HasFree(); } );
 	if ( Paged && !page )
 	{
@@ -7098,7 +7097,6 @@ void ObjectPool< T, N, Paged >::DeleteAll()
 	else
 	{
 		deleteAllFn( m_firstPage.Get() );
-		// m_pages.Append( m_firstPage.Get()->node );
 	}
 	m_length = 0;
 }
